requirements.txt
tweepy>=4.10.0,<5.0.0
google-generativeai>=0.4.0,<0.5.0
pandas>=1.0.0,<3.0.0
python-dotenv>=0.19.0,<1.0.0 # Good for local development if not using Codespaces, still useful for .env.example